+++ packages/brain-extraction/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @neurodesk/brain-extraction
 
+## 0.1.7
+
+### Patch Changes
+
+- @neurodesk/synthsr@0.4.20260923
+
 ## 0.1.6
 
 ### Patch Changes